requirements.txt
asgiref==3.6.0
certifi==2023.5.7
charset-normalizer==3.1.0
distlib==0.3.6
Django==4.2.3
django-cors-headers==3.14.0
django-filter==23.2
django-jazzmin==2.6.0
django-thread==0.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
filelock==3.12.0
idna==3.4
numpy==1.24.3
pandas==2.0.1
Pillow==9.5.0
platformdirs==3.5.1
Pygments==2.15.1
PyMySQL==1.0.3
python-dateutil==2.8.2
pytz==2023.3
requests==2.30.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.2
virtualenv==20.23.0
PyJWT==2.8.0
